requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
colorama==0.4.6; platform_system == 'Windows'
discord.py==2.3.2; python_full_version >= '3.8.0'
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
multidict==6.0.4; python_version >= '3.7'
openai==0.28.1; python_full_version >= '3.7.1'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
tqdm==4.66.1; python_version >= '3.7'
urllib3==2.0.6; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
